backend/scraper/requirements.txt
requests
beautifulsoup4
pytz
pymongo
python-dotenv